Add integration test for employeeService, fix bugs in employeeService…

+++ src/main/java/com/car/rental/employee/Employee.java
@@ -1,5 +1,6 @@
 package com.car.rental.employee;
 
+import java.math.BigDecimal;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -20,6 +21,6 @@ public class Employee {
     private String firstName;
     private String lastName;
     private String employmentPosition;
-    private Double salaryPerHour;
+    private BigDecimal salaryPerHour;
     private boolean deleted;
 }
